+++ rocAL_pybind/rocal_pybind.cpp
@@ -141,6 +141,9 @@ std::unordered_map<int, std::string> rocalToPybindLayout = {
     {1, "NCHW"},
     {2, "NFHWC"},
     {3, "NFCHW"},
+    {4, "NHW"},
+    {5, "NFT"},
+    {6, "NTF"}
 };
 
 std::unordered_map<int, std::string> rocalToPybindOutputDtype = {
@@ -379,6 +382,9 @@ PYBIND11_MODULE(rocal_pybind, m) {
         .value("NCHW", ROCAL_NCHW)
         .value("NFHWC", ROCAL_NFHWC)
         .value("NFCHW", ROCAL_NFCHW)
+        .value("NHW", ROCAL_NHW)
+        .value("NFT", ROCAL_NFT)
+        .value("NTF", ROCAL_NTF)
         .export_values();
     py::enum_<RocalDecodeDevice>(types_m, "RocalDecodeDevice", "Decode device type")
         .value("HARDWARE_DECODE", ROCAL_HW_DECODE)
@@ -722,6 +728,8 @@ PYBIND11_MODULE(rocal_pybind, m) {
     m.def("lensCorrection", &rocalLensCorrection,
           py::return_value_policy::reference);
     m.def("preEmphasisFilter", &rocalPreEmphasisFilter, 
+            py::return_value_policy::reference);
+    m.def("spectrogram", &rocalSpectrogram,
           py::return_value_policy::reference);
 }
 }  // namespace rocal